backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentflow"
version = "0.1.0"
description = "AgentFlow Article Publishing - single-user writing/publishing automation"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Mobius" }]
dependencies = [
    "anthropic>=0.40",
    "openai>=1.50",
    "tweepy>=4.14",
    "feedparser>=6.0",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "mistune>=3.0",
    "pyjwt>=2.8",
    "click>=8.1",
    "python-docx>=1.1",
    "trafilatura>=1.9",
    "requests>=2.31",
    "aiohttp>=3.9",
    "pydantic>=2.6",
    "python-multipart>=0.0",
    "Pillow>=10.0",
]

[project.scripts]
af = "agentflow.cli.commands:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentflow*", "api*"]

[tool.setuptools.package-data]
"agentflow.shared" = ["mocks/*.json", "mocks/*.txt"]
